Codereview changes. Updated format.
