--- src/main/java/marines/data/Item.java
@@ -1,35 +1,0 @@
-package main.java.marines.data;
-
-/**
- * @author Marines Lopez
- */
-public class Item {
-	private String name;
-	private int weight;
-
-	public Item(String name, int weight) {
-		this.name = name;
-		this.weight = weight;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public int getWeight() {
-		return weight;
-	}
-
-	public void setWeight(int weight) {
-		this.weight = weight;
-	}
-
-	@Override
-	public String toString() {
-		return "Item[name=" + name + ", weight=" + weight + "]";
-	}
-}
