+++ src/test/java/marines/SpaceShipTests.java
@@ -1,0 +1,24 @@
+package test.java.marines;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class SpaceShipTests {
+
+    @BeforeEach
+    void setUp() throws Exception {
+    }
+
+    @AfterEach
+    void tearDown() throws Exception {
+    }
+
+    @Test
+    void test() {
+        fail("Not yet implemented");
+    }
+
+}
